fix: remove KZG DA support in output deserialization
KZG DA when supported yields a different Starknet OS output,
which doesn't contain the state diff.

And the KZG proofs are also expected as input.This is currently not supported by piltover.
At the moment, since no work has been done to fully support KZG DA,
the support is disabled by panicking if the flag is set to true in the Starknet OS output.