fix: add validation of the bootloader hash in layout bridge output
Layout bridge is a verifier written in Cairo.
To ensure that the program that has been verified is legitimate,
and add consistency with the computed fact, the layout bridge output
that contains the executed program is now verified.

In the current implementation, this program hash is the bootloader hash,
since Starknet OS execution is bootloaded. And this is the trace
being verified by the layout bridge.